refactor: split resident menu presentation
